pyproject.toml
[project]
name = "linkedin-job-agent"
version = "0.1.0"
description = "Semi-automatic LinkedIn job hunting agent"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.7",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
